connecting the nav graph
