+++ .idea/misc.xml
@@ -8,24 +8,10 @@
         <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/calendar.xml" value="0.1145" />
         <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/ic_launcher_background.xml" value="0.1" />
         <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/profile_image.xml" value="0.1145" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/back_arrow.xml" value="0.1625" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/calendar.xml" value="0.1625" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/profile_image.xml" value="0.1625" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/rectangular_greyed.xml" value="0.231" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/rectangular_transparent.xml" value="0.231" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/splash_screen_gradient.xml" value="0.1" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/layout/activity_splash_screen.xml" value="0.22" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/layout/fragment_date_picker_dialog.xml" value="0.22447916666666667" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/layout/fragment_friend_profile.xml" value="0.22" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/layout/fragment_onboarding.xml" value="0.22447916666666667" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/layout/fragment_splash_screen.xml" value="0.22447916666666667" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/mipmap-anydpi-v26/ic_flower.xml" value="0.108" />
-        <entry key="..\:/Users/h/AndroidStudioProjects/WisherBirthdayWishApp/app/src/main/res/mipmap-anydpi-v26/ic_splash_screen.xml" value="0.1735" />
-        <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.1" />
-        <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/ic_launcher_background.xml" value="0.1" />
         <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/selected_background.xml" value="0.1" />
         <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/small_button_background.xml" value="0.1365" />
         <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/drawable/unselected_circle.xml" value="0.1" />
+        <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/layout/activity_base.xml" value="0.103125" />
         <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/layout/activity_splash_screen.xml" value="0.1" />
         <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/layout/fragment_friend_profile.xml" value="0.103125" />
         <entry key="..\:/Users/johnson.oyesina/StudioProjects/WisherBirthdayWishApp/app/src/main/res/layout/fragment_onboarding.xml" value="0.13632246376811594" />
