+++ app/src/main/java/com/itech/wisherbirthdaywishapp/views/ui/welcomeScreenFragment/WelcomeScreenFragment.kt
@@ -7,6 +7,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import androidx.navigation.fragment.findNavController
 import com.itech.wisherbirthdaywishapp.R
 import com.itech.wisherbirthdaywishapp.databinding.FragmentWelcomeScreenBinding
 import com.itech.wisherbirthdaywishapp.views.utils.TextFormatter
@@ -25,8 +26,12 @@ class WelcomeScreenFragment : Fragment(R.layout.fragment_welcome_screen) {
                 0, 5,
                 Typeface.BOLD
             )
-
-
+        binding.btnSignIn.setOnClickListener {
+            findNavController().navigate(R.id.action_welcomeScreenFragment_to_signIn)
+        }
+        binding.btnSignUp.setOnClickListener {
+            findNavController().navigate(R.id.action_welcomeScreenFragment_to_signUP)
+        }
 
     }
 }
