+++ app/src/main/res/layout/fragment_sign_u_p.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".signUP">
+    tools:context=".views.ui.loginFragment.signUP">
 
     <TextView
         android:layout_width="wrap_content"
