+++ app/src/main/res/layout/fragment_splash_screen.xml
@@ -4,6 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    tools:context=".views.ui.splashScreenFragment.SplashScreenFragment"
     android:background="@color/primaryDark">
 
     <ImageView
