+++ app/src/main/res/layout/fragment_welcome_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/purple_900"
-    tools:context=".views.WelcomeScreenActivity">
+    tools:context=".views.ui.welcomeScreenFragment.WelcomeScreenFragment">
 
 
     <TextView
